Clean up util functions and naming
